[REFACTOR] Deduplicate keyword logic
A number of keywords now share most of their implementation between the
Call and Append invocations, with a few minor differences. This PR
deduplicates the logic that they share into shared utility functions,
with a `toAppend` utility which wraps the Call version of the utility to
turn it into an Append version in a type-safe way.