Ensure paths off of keywords works in loose mode.
When in loose mode referencing `{{component.name}}` **must** mean
relying on implicit this fallback as `{{this.component.name}}`. The
existing code always assumes that any `PathExpression` _starting_ with a
keyword name (as its first path part) **must** be a keyword invocation.

This change ensures that `{{component.name}}` can properly fallback to
`{{this.component.name}}` when running in loose mode, and emits a
helpful error when running in strict mode.