[DOC lts] Remove mention on `transition-to` helper
The `transition-to` helper does not exists.

I suppose we could mention how to do this with the `{{on}}` modifier along with
a custom action, but that would still be inappropiate for accessibility reasons
in most cases. We should probably just deprecate this functionality, but in the
mean time, we can probably just leave it at this.