[DOC lts] Remove mention on `transition-to` helper
The `transition-to` helper does not exists.

I suppose we could mention how to do this with the `{{on}}` modifier along with
a custom action, but that would still be inappropiate for accessibility reasons
in most cases. We should probably just deprecate this functionality, but in the
mean time, we can probably just leave it at this.

(cherry picked from commit c5a0a594d6586f0ea4a4bae1fa97af7626c3c569)