[BUGFIX] Convert component names after symbol allocation
Currently component names are converted via the `customizeComponentName`
option prior to symbol allocation. This means that a component can end
up conflicting with an existing local variable, even though they
logically represent different things. This PR ensures that customization
happens after symbol allocation, as a final step within the symbol table
itself. This also means that the correct value is passed through the
rest of the AST and transformations, so it cannot conflict with keywords
as well as other language constructs.