Add `assert` and `deprecate` to global context API
This PR adds `assert` and `deprecate` to the global context API, and adds a new
package: `@glimmer/vm-babel-plugins`. This plugin exports a function that can
be used to generate Babel plugins which strip debug tooling in the VM. The idea
is that Ember/Glimmer.js will be able to re-export and use this in their own
Babel configs/utilities.