Update glimmer-vm to prevent errors for older inline precompilation
This updates glimmer-vm to avoid emitting the `scope` property on
precompiled template blocks when not using strict mode (when using
strict mode `scope` will either be `null` or a function).

This is specifically to work around a bug in older versions of
babel-plugin-htmlbars-inline-precompile (versions from v2.0.0 through
v4.0.1) where they would (incorrectly) attempt to build an
`ObjectExpression` when a `null` value is found in the template
precompilation result (D'oh!).

Note: those versions of babel-plugin-htmlbars-inline-precompile should
still be migrated away from (they are quite old, and will definitely not
work for newer features like strict mode).