Make defaultId function more robust
The current `defaultId` function attempts to load the `crypto` module
when in Node. However, it uses `require` directly, which may be
clobbered by AMD require (in the case of Ember). This was previously
handled by using Babel transforms, but that solution is somewhat
brittle. This updates the logic to check for the `module` global
instead, and defers to it if it exists, before checking for `require`.