fix: only return empty href when LinkTo href generation throws error
In 3.24.0-3.24.2 we made changes to let LinkTo href generation returns
empty when initial transition is not started. This is not true and
causing regression on the super-rental tutorial.

In previous tutorials, the LinkTo component will return empty href when
`this.owner.setupRouter()` is not called. It generates valid href if
`setupRouter()` is called and complete model are passed for dynamic
segments.

In our previous changes, we removed the requirement to call
`setupRouter`, and always return empty href when initial transition is
not started. This is not expected when passing complete dynamic segments
without initial transition.