[BUGFIX] Fixes dynamic helpers and modifiers with updating definitions
Currently, when the definition of a dynamic helper/modifier is updated,
the instance of that modifier or helper is not also updated. This was an
oversight from the initial implementation. This PR adds this
functionality, and also fixes a bug where sub-expressions could not be
the callee in an expression (e.g. {{(if @foo Bar) "baz"}}).

Notes:

Modifiers and helpers have been refactored to receive
CapturedArguments instead of VMArguments. In general, most helpers and
modifiers immediately capture the arguments anyways, so this shouldn't
affect performance. This is necessary because dynamic
helpers/modifiers must captured the arguments, otherwise they could
change, and if the helper or modifier is updated to a new definition,
it would get the wrong arguments.
Helpers no longer receive the VM itself. Instead, they receive the
owner they belong to, and optionally they receive dynamicScope.
This refactor unifies the way helpers are invoked in the VM and by
invokeHelper, which means invokeHelper can now be used with
built-in helpers such as array and hash.
Only static helpers (e.g. non-dynamic, non-invokeHelper) receive
dynamicScope. The only helper that needs this is {{outlet}}, which
does not make sense to invoke dynamically currently. We will assert
with a good error message in the implementation in Ember.
Helpers that wish to associate destroyables should now associate them
directly on the reference they return. This allows us to have a more
streamlined and unified API overall. Generally, only custom helpers
and a few select built-in helpers need this ability.
hasDestroyableChildren has been added in order to prevent
associating non-custom helpers if they don't have any children. This
should mean less overall memory usage/CPU cost in those cases, which
are likely the most common.
As an aside, this refactor does make me feel like even moreso that we
should rethink the way Helper is represented in the VM. The current
structure has evolved organically to handle custom helpers, but it feels
not ideal. The VM should likely be more aware of custom helpers in some
way, but that's a larger refactor.