[DEPRECATE] `registerPlugin` / `unregisterPlugin` and legacy class based AST plugins
The usage of `registerPlugin`/`unregisterPlugin` **requires** global
state mutation which leaks AST plugins between addon's and applications.
Tools like Embroider and ember-cli-htmlbars have avoided this by
ensuring that we purge the Node module caching system between each
attempt at using the template compiler. This results in _massive_ memory
growth (the entire `ember-template-compiler.js` source string is
included in memory for each addon in the build, as well as any JIT
artificats).

This is **very _bad_**.

This PR deprecates using `registerPlugin`/`unregisterPlugin` along with
the legacy AST transform API. All of these APIs are private APIs (as is
anything to do with the template compilation AST), but they are used
fairly often so a proper deprecation cycle is useful.