[BUGFIX] Do not eagerly consume modifier args during destruction
Currently, in the 3.13 modifier capabilities version, we eagerly consume
the arguments that are passed to the modifier when it is destroying.
This can cause issues with backtracking rerender when the state isn't
actually used, and neither is the consumption. This change makes it so
that we don't eagerly consume these arguments during destruction in
general.