[BUGFIX beta] Remove unnecessary module transform
The `injectNodeGlobals` transform would detect usage of `module` and
transform it to be compatible with Node. Recently the VM was updated to
make this unnecessary, and it turns out that the VM's more robust code
actually _breaks_ when used with this transform.

The transform was removed in a much larger refactor on master, which is
why it was not backported along with the other related fixes/changes.