[BUGFIX] Restore nested paths on run and computed
These values were previously nested on `run` and `computed`, and were
accessible from the _imported_ versions of those values due to the fact
that they transformed to be the same as globals. They were used by a few
different consumers, but were not intended to be used this way.

This PR restores the previous behavior, but also deprecates it.