[BUGFIX] Use getter for `currentRunLoop`
The `currentRunLoop` property on the global run was a getter, not a
value. I missed this when doing the conversion, and it doesn't appear
that we had a test for the behavior of that property, other than the
re-export test (which I also changed, mistakenly). This restores
the previous behavior, which deprecating it.