Change Typings for `setupController`
As `_transition` is not used, it should not be required. I'd prefer to take it out completely, but in that case, I'm not sure if there are going to be other ramifications. Happy to investigate that if there's appetite.

Also, happy to be told I'm totally off base here!

Cheers!