make transition optional on setupController [19501]
In issue 19501 it was noted that transition was marked as required
on the API for the setupController (and would fail for Typescript
projects if not present).

This change makes transition optional as it is unused by
setupController and should therefore not break applications which
do not pass it in.

Removing the parameter would potentially break backwards
compatibility. As a compromise, it's marked as optional only.