Simplify stack
Simplifies the VM stack, removing the WasmStack and InnerStack in favor
of a single stack. This stack no longer encodes values, so we don't need
to maintain separate JS and handle stacks. Overall this simplifies the
architecture and makes it easier to understand and maintain.

In the future, when we do look into introducing WASM again, we can
bring this separation of concerns back at that time.