[BUGFIX beta] Remove Namespace processing
Ember no longer uses Namespaces in the `toString` on classes, so there's
no reason to eagerly process namespaces on application boot. This is
currently triggering the deprecation message in new Ember applications
on Beta, so this removes it to prevent that.

(cherry picked from commit e1323e283b0212ee037437722b4ebf98c67f8587)