Allow all (hash) keys to be set
Enables setting all keys on a `(hash)`. We do this by creating a
side-effecting cache which syncs the value of the key to the local state
whenever it is first accessed, and checking that cache on all subsequent
accesses. Whenever the upstream value has dirtied, the cache will rerun
and overwrite the local state of the hash.

When setting values on the hash with `Ember.get` and `Ember.set`, these
functions will get and set the tags associated with the _proxy_ for
those keys. This is why `customTagFor` has been updated to get and
combine the tag for the proxy as well, for the time being.