Strip assert and deprecate from the benchmark
When we added the `assert` and `deprecate` global context APIs, we did
not add removal of them to the benchmark's build process. This PR adds
`@rollup/plugin-strip` to handle removing them. It appears to work well,
as long as the names are consistent and not overlapping (which they are
currently).