Remove UpdatingOpcode base classes
UpdatingOpcodes currently extend from two levels of abstract classes,
which are mostly unnecessary. They only add a `guid` field which is no
longer used, and a `type` field also appears to not be used. This PR
removes the abstract classes and extra fields, and replaces them with
an interface.

In addition this PR removes some unused data structures and utilities
that were related (such as the GUID utils).