[BIGFIX release] Fix `<LinkTo>` with nested children
During bubbling, `event.target` may point to a child element whereas
`event.currentTarget` always points to the element where the handler
was attached, which is what we want here.

Reported in a comment on #19546, though this may be a distinct issue
from the original report as it was reported as a default-cancelling
parent element interfering with the nested `<LinkTo>`, and this is
the other way around.

(cherry picked from commit 57907a4d7d91aa58b3b3ac1258e639ce221f9492)