Revert hash to bi-modal behavior
This PR reverts hash to it's bi-modal behavior prior to Ember ~3.22.
Hash will now:

1. Eagerly produce an object whenever used in JS, which consumes all
   references and autotracks them. Whenever one reference changes, the
   hash will dirty and produce a new object.
2. Lazily allow access to child paths in templates, e.g. `{{hash.foo}}`
   will _not_ access any other references on the hash.

These behaviors have some quirks, but the quirks existed before 3.22 and
were primarily related to the fact that the object produced in 1 is
mutable and can diverge from the values in 2. Since setting values on a
hash object is deprecated, this should not be an issue in the near
future.