drop GetFreeAsFallback
GetFreeAsFallback got broken by 179a388d69af6fcd1bc906de396bde55e4925933

As far as I can tell, it's never correct to call `withPath` by itself, it's always supposed to come after an opcode that leaves a value on the stack that needs path support.

Rather than fix it, I'm going to try to cut the rest of it out, since it exists only to support `this` fallback.