drop GetFreeAsFallback
GetFreeAsFallback got broken by ab0d51fe8bae69eec605632338ac9b3558cc9ea9

As far as I can tell, it's never correct to call `withPath` by itself, it's always supposed to come after an opcode that leaves a value on the stack that needs path support.

Rather than fix it, I'm going to try to cut the rest of it out, since it exists only to support `this` fallback.