Improve implicit injections deprecation for routes
The store property on routes uses a setter, and so was not impacted by
the implicit-injections deprecation issued elsewhere. This may lead app
or addon authors to miss usage of a deprecated store injection on the
road to 4.0.

This improvement of the deprecation fidelity means libraries which use
the type injection system to define a store, like Ember Data, should
consider another approach to maintain any undeprecated API (for example
reopening the route class).