[BUGFIX release-3-28] Fix #19867
The way most Ember apps are setup in late 3.28 (an implicit injection of
`store` but also an explicit injection) was not under test. This patch
adds a test, and fixes a bug when applications use that pattern.

The prior implementation of the `store` injection deprecation on routes
(added in https://github.com/emberjs/ember.js/pull/19854) used a wrapper
`DeprecatedStoreInjection` in some cases which was unwrapped by the
store getter. A deprecation implemented on `CoreObject` for when a
explicit and implicit deprecation mis-match conflicted with that logic.

Here, refactor away from the `DeprecatedStoreInjection` wrapper to
instead use a WeakSet to track injected instances.