[BUGFIX] Improve types for DeprecationOptions
The previous design allowed invalid values for `DeprecationOptions`:

- `{}` (that is, an empty object)
- `{ enabled: 'v4.1.0' }` (that is, without `available`)

Both of these violate the text of the RFC! The new design *requires*
that at least `available` be present, and that `enabled` cannot be
present *without` `enabled`. (Previously, a number of Ember's own
internal deprecations were not following the RFC spec; the newly
introduced types caught all of those!)

This PR also removes the exported key names for the stages, which are
implicit in the interface-based API and which are currently unused
outside this module.

While touching the code anyway, fix a bunch of missing return type
annotation lint errors.