Document `router.on()` and `router.off()` methods
The methods to register/unregister listeners for the route change events on the router service are not documented as of now.
I also updated the example for the events to not register the events in the route constructor, as I guess the actual example there would rather use the activate hook or something like this.

Related to https://github.com/typed-ember/ember-cli-typescript/issues/1161