Allow components to be named the same as void tags (#1390)
* Printer drops closing tag for components with names similar to standard HTML tags

This includes a failing test case. If you parse to ast and immediately print back out, it turns:

```hbs
<Link></Link>
```

into

```hbs
<Link>
```

This behavior is special for `<Link>`. Other names don't do this.

I suspect it's because, in strict HTML spec terms `<Link>` is the same as `<link>`, and `<link>` is supposed to not have a closing tag. But in Glimmer/Ember terms, `<Link>` is a component and dropping the closing tag introduces a syntax error.

This was found in https://github.com/embroider-build/embroider/issues/1176

* Do not call .toLowerCase() on tag names, so that we can have components named the same as void elements

* Test all the potential void tags

* Add isVoidTag to help with additional void tag checks and improve readability of the Printer

Co-authored-by: Edward Faulkner <edward@eaf4.com>