[internal] improve Cache TS and implementation
I happened to look at this while working on publishing preview types,
and noticed that the types here were using a non-null assertion, and
switched it to an explicit cast with a `// SAFETY: ...` comment.

Next, use the optional parameter initialization shorthand to make the
constructor more clear (a constructor which defines properties *and*
has a body is extremely confusing to contributors, including me!) and
avoid doing something in the body which the language can do for us.

Finally, switch one use of `any` to `unknown`, in one of the usage sites
for `Cache`.