[FEATURE] Test all supported TS versions in CI
As part of implementing RFC 0800, introduce support for testing all
supported versions of TypeScript against our types. For the moment, the
only additional supported version is the TypeScript nightly (`next`)
build. When we get to publishing stable types, the current version will
be the first version we support, and we will add new versions to the
matrix over time in line with the "rolling window" support policy.

Additionally, *remove* type checking from the `lint` scripts and create
a new set of `type-check` scripts instead. This is pragmatic, not
philosophical: we treat it differently from the "lints" here in that we
need to be able to run it repeatedly against different versions,
whereas the lints are all one-and-done.