Include package name in deprecation function
So this slightly changes the deprecation message to allow for package (namespace) to be included in the error message. This make it so that the message "will be removed in ember (package version)" is replaced with "will be removed in (package name) (package version)".

This was change was proposed because of this usage:
https://github.com/adopted-ember-addons/ember-metrics/blob/f604272de70e44ac731c711cfad5eae8792ee3ca/addon/metrics-adapters/google-tag-manager.js#L50

and this issue:
https://github.com/adopted-ember-addons/ember-metrics/issues/453

(cherry picked from commit 811fd074c82943b8bb33455ada15e58f034f8736)