Prove out basic stable type generation
This isn't remotely ready for primetime, but it does prove out the
basic viability of @chadhietala's suggestion that we may not need to
fix all the circularities before we can publish. We in fact can just
publish as is: none of the circularities are *hard* blockers, at least
for *this* part of the effort (though they will be for getting docs
published with these as the source of truth).

The things we will need to do to be able to execute the rest of the way
on this, given our goals:

- Fix a couple cases where we are using private names in public types,
  specifically around the `OWNER` from `@glimmer/owner`.

- Put the hacky `generate-tsconfigs.mjs` script somewhere besides the
  root, expand its capabilities to include wrapping generated modules
  in a `declare module` statement and running Prettier on the result,
  and rename it accordingly.

- Land @wagenet's in-progress PR (#20175) to convert the `ember`
  package to TS so we can publish types for it, which will unblock some
  of the other Ember packages as well.

- Properly exclude the parts of Ember's APIs we *don't* want to be
  publishing (all the purely-internal and private stuff).