Revert to defineProperty for get/set on Ember namespace
While we might be able to make the change over to using native getters
and setters in a future PR, we want this change to be exactly and only
a type conversion, exactly preserving the previous behavior. Switching
back to using `Object.defineProperty()` for the getters and setters
defined on the `Ember` namespace preserves the previous behavior, which
our tests correctly caught for us!