Fix type definition of `getOwner`, and explain it
Use `GlimmerComponent<any>` instead of plain `GlimmerComponent`, since
the latter resolves to the default `GlimmerComponent<unknown>`, which
ends up being too narrow to resolve correctly with `getOwner` when any
actual signature is applied.

As a bonus, rename `FrameworkObject` in this context to the somewhat
more verbose `KnownFrameworkObject`, which avoids a conflict with the
actual `FrameworkObject` type used internally.