Move docs for `getOwner` to `@ember/owner`
This is a TSDoc-friendly/future-proofing change: most of the items which
`@ember/owner` re-exports from `@ember/-internals/owner` can be
documented at the definition site, because they are plain re-exports.
`getOwner` is a bit different, because (as discussed in an earlier
commit) the public `Owner` type it returns has much less detail than our
`InternalOwner` type. We do this via a reassignment of the item, so the
docs need to move to that reassignment to be visible for TS consumers
and, in the future, for TSDoc extraction of docs.

The existing YUIDoc infrastructure will be unaffected by this move.