Make intern utility type safe
Experimenting with removing/improving some *other* types while
implementing RFC 0821. revealed that this function needed to preserve
its input type when that type was narrower than just `string`, so
introduce a type parameter to keep that information around. Then,
switch to using `Object.create(null)` so it can be used in the `Record`
type safely.

Pulling this out into its own commit makes it easier to land, and it
also allows us to guarantee that we are making changes *only* to the
types and module layout for the PR implementing RFC 0821.